+++ libraries/stdlib/unsigned/src/kotlin/UShort.kt
@@ -42,14 +42,14 @@ public inline class UShort internal constructor(private val data: Short) : Compa
      * Returns zero if this value is equal to the specified other value, a negative number if it's less than other,
      * or a positive number if it's greater than other.
      */
-    public operator fun compareTo(other: UInt): Int = this.toUInt().compareTo(other.toUInt())
+    public operator fun compareTo(other: UInt): Int = this.toUInt().compareTo(other)
 
     /**
      * Compares this value with the specified value for order.
      * Returns zero if this value is equal to the specified other value, a negative number if it's less than other,
      * or a positive number if it's greater than other.
      */
-    public operator fun compareTo(other: ULong): Int = this.toULong().compareTo(other.toULong())
+    public operator fun compareTo(other: ULong): Int = this.toULong().compareTo(other)
 
     /** Adds the other value to this value. */
     public operator fun plus(other: UByte): UInt = this.toUInt().plus(other.toUInt())
